backend/pyproject.toml
[project]
name = "fastapi-mongodb-blog-backend"
version = "0.1.0"
description = "Educational project on developing a blog using Fast API, MongoDB, and other technologies."
license = {text = "MIT"}
authors = [
    {name = "desunovu", email = "desunovu@gmail.com"},
    {name = "bailu", email = "weiqihuang000@gmail.com"}
]
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "uvicorn>=0.23.2",
    "beanie>=1.30.0",
    "email-validator>=2.0.0.post2",
    "fastapi[standard]>=0.129.1",
    "python-multipart>=0.0.6",
    "bcrypt>=4.0.1",
    "python-jose>=3.3.0",
    "cryptography>=41.0.3",
    "openai>=0.28.1",
    "pyyaml>=6.0.1",
    "langdetect>=1.0.9",
]

[dependency-groups]
dev = [
    "black>=23.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["blogapp"]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
